Feature #70119 VSChartAxesVisibilityController service proxy
